requirements.txt
Babel==2.5.3
beautifulsoup4==4.6.0
blinker==1.4
certifi==2018.4.16
chardet==3.0.4
click==6.7
cssselect==1.0.3
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.0.2
Flask-Babel==0.11.2
Flask-Navigation==0.2.0
Flask-PyMongo==0.5.1
Flask-Table==0.5.0
Flask-WTF==0.14.2
idna==2.6
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.10
lxml==4.2.1
MarkupSafe==1.0
mccabe==0.6.1
newspaper3k==0.2.6
nltk==3.2.5
Pillow==5.1.0
pycodestyle==2.3.1
pyflakes==1.6.0
pymongo==3.6.1
python-dateutil==2.7.2
pytz==2018.4
PyYAML==3.12
requests==2.18.4
requests-file==1.4.3
six==1.11.0
tldextract==2.2.0
tzlocal==1.5.1
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.1
